requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.2
frozenlist==1.8.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
soupsieve==2.8
tqdm==4.67.1
typing_extensions==4.15.0
yarl==1.22.0
PyYAML==6.0.3
playwright==1.47.0
brotli
requests
bs4
Flask==3.0.3

# Django and REST Framework
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
drf-spectacular==0.29.0

# Celery and Redis
celery==5.6.0
redis==7.1.0

# External tools (install manually if not present):
# subfinder
# httpx
# nuclei
